requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
behave==1.2.6
behave-django==1.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1
django-cors-headers==3.13.0
django-import-export==2.8.0
django-oauth-toolkit==2.1.0
django-rest-swagger==2.2.0
django-silk==0.5.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.8.1
drf-yasg==1.21.3
et-xmlfile==1.1.0
factory-boy==3.2.1
Faker==13.15.1
h11==0.13.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.9.1
jwcrypto==1.3.1
MarkupPy==1.14
MarkupSafe==2.1.1
oauthlib==3.2.0
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.4.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tablib==3.2.1
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
